29/11/2025 profile done

+++ app/src/main/java/com/bos/payment/appName/data/model/justpaymodel/GetToselfPayoutCommercialReq.kt
@@ -5,10 +5,10 @@ import com.google.gson.annotations.SerializedName
 data class GetToselfPayoutCommercialReq(
 
 	@field:SerializedName("txtslabamtfrom")
-	val txtslabamtfrom: Int? = null,
+	val txtslabamtfrom: Double? = null,
 
 	@field:SerializedName("txtslabamtto")
-	val txtslabamtto: Int? = null,
+	val txtslabamtto: Double? = null,
 
 	@field:SerializedName("merchant")
 	val merchant: String? = null,
