+++ app/src/main/java/com/bos/payment/appName/data/model/merchant/apiServiceCharge/mobileCharge/GetCommercialReq.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 
 data class GetCommercialReq (
 
-  @SerializedName("txtslabamtfrom"  ) var txtslabamtfrom  : Int?    = null,
-  @SerializedName("txtslabamtto"    ) var txtslabamtto    : Int?    = null,
+  @SerializedName("txtslabamtfrom"  ) var txtslabamtfrom  : Double?    = null,
+  @SerializedName("txtslabamtto"    ) var txtslabamtto    : Double?    = null,
   @SerializedName("merchant"        ) var merchant        : String? = null,
   @SerializedName("productId"       ) var productId       : String? = null,
   @SerializedName("cantentType"     ) var cantentType     : String? = null,
